Revert incorrect fix compilation on Mac. This reverts commit ef6b20030cc6f38e3d0ba340a569d934f246f616.

+++ dbms/src/Columns/ColumnDecimal.h
@@ -166,7 +166,7 @@ template <typename T>
 template <typename Type>
 ColumnPtr ColumnDecimal<T>::indexImpl(const PaddedPODArray<Type> & indexes, UInt64 limit) const
 {
-    UInt64 size = indexes.size();
+    size_t size = indexes.size();
 
     if (limit == 0)
         limit = size;
